:hammer: Extract notify orchestration
