+++ src/SFA.DAS.Zendesk.Monitor.Function/ZenWatchDurableFunction.cs
@@ -8,41 +8,9 @@ using System.Threading.Tasks;
 
 namespace ZenWatchFunction
 {
-    internal class NotifyTicket
-    {
-        public long Id { get; set; }
-    }
-
-    internal class NotifyTicketValidator : AbstractValidator<NotifyTicket>
-    {
-        public NotifyTicketValidator()
-        {
-            RuleFor(x => x.Id).NotEmpty();
-        }
-    }
-
-    public class WatcherOrchestration
+     public class WatcherOrchestration
     {
         private static readonly RetryOptions retry = new RetryOptions(TimeSpan.FromSeconds(1), 5);
-
-        [FunctionName("NotifyTicket")]
-        public static async Task<HttpResponseMessage> HttpStart(
-            [HttpTrigger(AuthorizationLevel.Anonymous, "get", "post")]HttpRequestMessage request,
-            [OrchestrationClient]DurableOrchestrationClient starter,
-            ILogger log)
-        {
-            var input = await request.GetJsonBody<NotifyTicket, NotifyTicketValidator>();
-
-            return await input.Match(
-                valid => StartNotifyTicket(valid),
-                invalid => Task.FromResult(request.BadRequest(invalid)));
-
-            async Task<HttpResponseMessage> StartNotifyTicket(NotifyTicket ticket)
-            {
-                var instanceId = await starter.StartNewAsync(nameof(ShareListedTickets), new[] { ticket.Id });
-                return starter.CreateCheckStatusResponse(request, instanceId);
-            }
-        }
 
         [FunctionName(nameof(ShareAllTickets))]
         public static async Task ShareAllTickets([OrchestrationTrigger] DurableOrchestrationContext context)
